# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Informasi Manajemen Inventaris Sekolah API",
        "description": "REST API untuk Sistem Informasi Manajemen Inventaris Sekolah. Mendukung modul Auth, Pengguna, Peran & Akses, Master Sekolah, Master Barang, Manajemen Aset, Transaksi Peminjaman & Permintaan, Pemeliharaan & Akhir Siklus Aset, serta Manajemen Database.",
        "contact": {
            "name": "Admin Sekolah",
            "email": "admin@sekolah.sch.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/database/reset": {
            "post": {
                "tags": [
                    "Database"
                ],
                "summary": "Reset database",
                "description": "Menghapus semua tabel dan menjalankan ulang migrasi serta seeder. PERHATIAN: Semua data akan hilang!",
                "operationId": "resetDatabase",
                "responses": {
                    "200": {
                        "description": "Database berhasil direset",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DatabaseResetResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Gagal mereset database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DatabaseErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/database/backup": {
            "post": {
                "tags": [
                    "Database"
                ],
                "summary": "Backup database",
                "description": "Membuat file backup SQL dari database saat ini dan mengembalikannya sebagai file download.",
                "operationId": "backupDatabase",
                "responses": {
                    "200": {
                        "description": "File backup SQL berhasil diunduh",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Gagal membuat backup database"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/database/restore": {
            "post": {
                "tags": [
                    "Database"
                ],
                "summary": "Restore database",
                "description": "Mengembalikan database dari file SQL yang diunggah. PERHATIAN: Data saat ini akan ditimpa!",
                "operationId": "restoreDatabase",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "sql_file"
                                ],
                                "properties": {
                                    "sql_file": {
                                        "description": "File SQL untuk restore",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Database berhasil direstore",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DatabaseRestoreResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Gagal merestore database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DatabaseErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/database/change-connection": {
            "post": {
                "tags": [
                    "Database"
                ],
                "summary": "Ubah koneksi database",
                "description": "Mengubah konfigurasi koneksi database di file .env.",
                "operationId": "changeConnectionDatabase",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DatabaseChangeConnectionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Koneksi berhasil diubah",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DatabaseChangeConnectionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DatabaseResetResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Database berhasil direset."
                    }
                },
                "type": "object"
            },
            "DatabaseRestoreResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Database berhasil direstore."
                    }
                },
                "type": "object"
            },
            "DatabaseChangeConnectionRequest": {
                "required": [
                    "db_host",
                    "db_name",
                    "db_user"
                ],
                "properties": {
                    "db_host": {
                        "type": "string",
                        "example": "127.0.0.1"
                    },
                    "db_name": {
                        "type": "string",
                        "example": "inventaris_sekolah_db"
                    },
                    "db_user": {
                        "type": "string",
                        "example": "root"
                    },
                    "db_pass": {
                        "type": "string",
                        "example": "secret123",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "DatabaseChangeConnectionResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Koneksi berhasil diubah."
                    }
                },
                "type": "object"
            },
            "DatabaseErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Error message"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}